moving tests around; added tests for projects methods
