+++ test/integration/v1/basic.js
@@ -1,0 +1,40 @@
+var expect = require( "chai" ).expect,
+    request = require( "supertest" ),
+    iNaturalistAPI = require( "../../../lib/inaturalist_api" ),
+    app = iNaturalistAPI.server( );
+
+describe( "routes", function( ) {
+
+  describe( "index", function( ) {
+    it( "redirects to /v1/docs", function( done ) {
+      request( app ).get( "/" ).
+        expect( "Location", "/v1/docs", done);
+    });
+  });
+
+  describe( "docs", function( ) {
+    it( "redirects to /v1/docs", function( done ) {
+      request( app ).get( "/docs" ).
+        expect( "Location", "/v1/docs", done);
+    });
+  });
+
+  describe( "swaggerRedirect", function( ) {
+    it( "redirects to /v1/swagger.json", function( done ) {
+      request( app ).get( "/swagger.json" ).
+        expect( "Location", "/v1/swagger.json", done);
+    });
+  });
+
+  describe( "swaggerJSON", function( ) {
+    it( "renders the swagger JSON file", function( done ) {
+      request( app ).get( "/v1/swagger.json" ).
+        expect( function( res ) {
+          expect( res.body.swagger ).to.eq( "2.0" );
+          expect( res.body.info.title ).to.eq( "iNaturalist API" );
+        }).
+        expect( "Content-Type", /json/ ).expect( 200, done );
+    });
+  });
+
+});
